+++ lib/src/realm_class.dart
@@ -483,6 +483,16 @@ class Realm implements Finalizable {
       throw RealmError('Starting a write transaction on a frozen Realm is not allowed.');
     }
   }
+
+  /// Compacts a Realm file. A Realm file usually contains free/unused space. 
+  /// 
+  /// This method removes this free space and the file size is thereby reduced. 
+  /// Objects within the Realm file are untouched.
+  /// Note: The file system should have free space for at least a copy of the Realm file. This method must not be called inside a transaction. 
+  /// The Realm file is left untouched if any file operation fails.
+  bool compact() {
+    return realmCore.compact(this);
+  }
 }
 
 /// Provides a scope to safely write data to a [Realm]. Can be created using [Realm.beginWrite] or